added view configuration loading/saving
Pressing Ctrl+S or F6 saves the current view configuration
(mode, zoom level, relative position, scroll speed etc.)
in a "sidecar" file with the name name and path as the
current image, but an additional ".pxv" extension.

On loading an image, the view configuration is restored
from the .pxv file, if present.
Syntax errors in that file are silently ignored
(except for console messages in debug builds).