ANSI: SAUCE parser and importer
If there's SAUCE, we can auto-configure the ANSI options from that now.

Side effect: The ANSI options are reset to their defaults when reloading
or changing files now. This makes the behavior more consistent with the
other view settings, and it eliminates issues when switching from a
SAUCEd Amiga ANSI to a non-SAUCEd PC ANSI or vice-versa.