fix ANSI rendering with PabloDraw 24-bit color sequences
Turns out that AnsiLove simply assumes that GD's internal pixel format
for TrueColor images is BGRX, instead of going through
gdImageColorAllocate() like a good citizen.

Fortunately, the GL_BGRA pixel format is a thing and we don't care
about the alpha channel anyway, so it's an easy thing to solve.